+++ src/trading/orders/order_helper.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Callable
 
 from api.interfaces.order import Order
@@ -7,6 +8,6 @@ class OrderHelper:
     @staticmethod
     def less_than_price_filter(price: str) -> Callable:
         def price_filter(order: Order) -> bool:
-            return float(price) > float(order.price)
+            return Decimal(price) > order.price
 
         return price_filter
